fix: unblock check.build.ui and check.lint.ui
Extract Configure enter session helper under eslint line budget;
fix typed canvas spec fixture in unit test.
