+++ web_src/.eslint-budget-baseline.json
@@ -15,5 +15,5 @@
     "max-lines": 3668,
     "complexity": 198
   },
-  "updatedAt": "2026-08-12T12:23:31.749Z"
+  "updatedAt": "2026-08-12T12:42:43.213Z"
 }
