+++ web_src/src/pages/app/useFactoryConfigureEnter.ts
@@ -1,6 +1,8 @@
 import type { CanvasesCanvas, CanvasesCanvasVersion } from "@/api-client";
 import { useEffect, useRef, type Dispatch, type MutableRefObject, type SetStateAction } from "react";
 
+import { startFactoryConfigureEnter, type FactoryConfigureEnterDeps } from "./factoryConfigureEnterSession";
+
 type UseFactoryConfigureEnterOptions = {
   factoryConfigure: boolean;
   editSessionActive: boolean;
@@ -12,17 +14,10 @@ type UseFactoryConfigureEnterOptions = {
   liveCanvasVersion?: CanvasesCanvasVersion | null;
   liveCanvas?: CanvasesCanvas | null;
   previewingCurrentVersionRef: MutableRefObject<boolean>;
-  activateCanvasVersionForEditing: (
-    versionId: string,
-    version: CanvasesCanvasVersion,
-    options?: { preserveStagedLayer?: boolean },
-  ) => void;
+  activateCanvasVersionForEditing: FactoryConfigureEnterDeps["activateCanvasVersionForEditing"];
   draftCanvasSpecsRef: MutableRefObject<Map<string, CanvasesCanvas["spec"] | null>>;
   setDraftCanvasSpec: Dispatch<SetStateAction<CanvasesCanvas["spec"] | null>>;
-  resyncStagedEditorState: (
-    versionId: string,
-    options?: { bumpResetNonce?: boolean; preferCachedStagedSpec?: boolean },
-  ) => Promise<void>;
+  resyncStagedEditorState: FactoryConfigureEnterDeps["resyncStagedEditorState"];
   setLastSavedWorkflowSnapshot: (workflow: CanvasesCanvas | null) => void;
 };
 
@@ -51,7 +46,7 @@ export function useFactoryConfigureEnter({
   liveCanvasRef.current = liveCanvas;
   const liveCanvasVersionRef = useRef(liveCanvasVersion);
   liveCanvasVersionRef.current = liveCanvasVersion;
-  const enterDepsRef = useRef({
+  const enterDepsRef = useRef<FactoryConfigureEnterDeps>({
     activateCanvasVersionForEditing,
     draftCanvasSpecsRef,
     previewingCurrentVersionRef,
@@ -110,62 +105,16 @@ export function useFactoryConfigureEnter({
     }
 
     inFlightVisitIdRef.current = visitId;
-    // Seed draft from the live version directly — do not wait on the versions
-    // list (handleUseVersion), which often left Configure with no activeVersionId.
-    const immediateSpec = liveCanvasRef.current?.spec ?? version.spec ?? { nodes: [], edges: [] };
-    const versionForEdit: CanvasesCanvasVersion = {
-      ...version,
-      metadata: {
-        ...version.metadata,
-        id: liveCanvasVersionId,
-      },
-      spec: immediateSpec,
-    };
-    const configureVersionId = liveCanvasVersionId;
-    let cancelled = false;
-    const {
-      activateCanvasVersionForEditing: activate,
-      draftCanvasSpecsRef: specsRef,
-      previewingCurrentVersionRef: previewingRef,
-      resyncStagedEditorState: resync,
-      setDraftCanvasSpec: setDraft,
-      setEditSessionActive: setEditActive,
-      setLastSavedWorkflowSnapshot: setSnapshot,
-    } = enterDepsRef.current;
-
-    previewingRef.current = true;
-    activate(configureVersionId, versionForEdit, { preserveStagedLayer: true });
-    specsRef.current.set(configureVersionId, immediateSpec);
-    setDraft(immediateSpec);
-
-    // Await staged resync before enabling edit so a late applyStagedSpec cannot
-    // wipe edits typed against the immediate seed.
-    void (async () => {
-      try {
-        await resync(configureVersionId, { bumpResetNonce: false });
-      } catch {
-        // Keep the immediate live/committed spec so Configure stays usable.
-      }
-      if (cancelled || configureVisitIdRef.current !== visitId) {
-        return;
-      }
-      editEnabledVisitIdRef.current = visitId;
-      inFlightVisitIdRef.current = null;
-      setEditActive(true);
-      const canvas = liveCanvasRef.current;
-      if (canvas) {
-        const spec = specsRef.current.get(configureVersionId) ?? immediateSpec;
-        setSnapshot({ ...canvas, spec });
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-      // Strict Mode remount: allow retry only when edit never enabled for visit.
-      if (editEnabledVisitIdRef.current !== visitId && inFlightVisitIdRef.current === visitId) {
-        inFlightVisitIdRef.current = null;
-      }
-    };
+    return startFactoryConfigureEnter({
+      visitId,
+      configureVersionId: liveCanvasVersionId,
+      version,
+      liveCanvasRef,
+      configureVisitIdRef,
+      inFlightVisitIdRef,
+      editEnabledVisitIdRef,
+      deps: enterDepsRef.current,
+    });
   }, [
     canStageCanvasVersion,
     canvasLoading,
